% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataclass-package.R
\docType{package}
\name{dataclass-package}
\alias{dataclass-package}
\title{dataclass: Easily Create Structured Lists or Data Frames with Input Validation}
\description{
Easily define templates for lists and data frames that validate each element. Specify the expected type (i.e., character, numeric, etc), expected length, minimum and maximum values, allowable values, and more for each element in your data. Decide whether violations of these expectations should throw an error or a warning. This package is useful for validating data within R processes which pull from dynamic data sources such as databases and web APIs to provide an extra layer of validation around input and output data.
}
\author{
\strong{Maintainer}: Chris Walker \email{walkerjameschris@gmail.com} [copyright holder]

}
\keyword{internal}
