% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{print.dataclass}
\alias{print.dataclass}
\title{Defines print behavior for dataclass objects}
\usage{
\method{print}{dataclass}(x)
}
\arguments{
\item{x}{A dataclass object}
}
\value{
A function with the following properties:

* An argument for each element provided to dataclass()
* Each argument in the returned function will validate inputs
* An error occurs if new elements passed to the returned function are invalid
* List is returned if new elements passed to the returned function are valid
}
\description{
This function serves to define print behavior for dataclass objects and does
not need to be called directly.
}
\examples{
# Define a dataclass for creating data! Pass to data_validator():
my_df_dataclass <-
  dataclass(
    dte_col = dte_vec(),
    chr_col = chr_vec(),
    # Custom column validator ensures values are positive!
    new_col = function(x) all(x > 0)
  ) |>
  data_validator()

# Nicely prints out required elements
my_df_dataclass
}
