% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\name{data_validator}
\alias{data_validator}
\title{Convert a dataclass to a data frame validator}
\usage{
data_validator(x)
}
\arguments{
\item{x}{A dataclass object}
}
\description{
If you intend to use your dataclass to validate data frame like object such
as tibbles, data frames, or data tables, pass the dataclass into this 
function to modify behavior.
}
\examples{
\dontrun{
my_df_dataclass <-
 dataclass(
   dte_col = dte_vec(),
   chr_col = chr_vec(),
   # Custom column validator which ensures column is numeric and postitive!
   new_col = function(x) num_vec(x) && all(x > 0)
 ) \%>\%
 # Tells dataclass it will be used on tibbles or data frames
 data_validator()

# Validate a tibble!
tibble(
 dte_col = as.Date("2022-01-01"),
 chr_col = "String!",
 new_col = 100
) \%>\%
 my_df_dataclass()
}
}
